[MNT]Moves DAG.fit method to DiscreteBayesianNetwork.fit (#2341)
### Your checklist for this pull request
Please review the [guidelines for contributing](CONTRIBUTING.md) to this
repository.

- [ ] Make sure you are requesting to **pull a topic/feature/bugfix
branch** (right side). Don't request your master!
- [ ] Make sure you are making a pull request against the **dev branch**
(left side). Also you should start *your branch* off *our dev*.
- [ ] Check the commit's or even all commits' message styles matches our
requested structure.

### Issue number(s) that this pull request fixes
- Fixes #

### List of changes to the codebase in this pull request
- 
-
-
