+++ pgmpy/tests/test_models/test_FunctionalBayesianNetwork.py
@@ -4,12 +4,6 @@ import numpy as np
 import pandas as pd
 from skbase.utils.dependencies import _check_soft_dependencies
 
-from pgmpy.utils._safe_import import _safe_import
-
-pyro = _safe_import("pyro", pkg_name="pyro-ppl")
-dist = _safe_import("pyro.distributions", pkg_name="pyro-ppl")
-torch = _safe_import("torch")
-
 from pgmpy import config
 from pgmpy.factors.continuous import LinearGaussianCPD
 from pgmpy.factors.discrete import TabularCPD
@@ -18,6 +12,8 @@ from pgmpy.models import FunctionalBayesianNetwork, LinearGaussianBayesianNetwor
 from pgmpy.utils import get_example_model
 from pgmpy.utils._safe_import import _safe_import
 
+pyro = _safe_import("pyro", pkg_name="pyro-ppl")
+dist = _safe_import("pyro.distributions", pkg_name="pyro-ppl")
 torch = _safe_import("torch")
 
 
@@ -237,7 +233,7 @@ class TestFBNMethods(unittest.TestCase):
 
         params = model.fit(
             data,
-            method="svi",
+            estimator="svi",
             optimizer=pyro.optim.Adam({"lr": 0.05}),
             seed=42,
             num_steps=100,
@@ -293,7 +289,7 @@ class TestFBNMethods(unittest.TestCase):
 
         params = model.fit(
             data,
-            method="SVI",
+            estimator="SVI",
             optimizer=pyro.optim.Adam({"lr": 0.05}),
             seed=42,
             num_steps=100,
@@ -357,7 +353,7 @@ class TestFBNMethods(unittest.TestCase):
         pyro.clear_param_store()
         params = model.fit(
             data,
-            method="MCMC",
+            estimator="MCMC",
             prior_fn=prior_fn,
             seed=42,
             num_steps=100,
@@ -422,7 +418,7 @@ class TestFBNMethods(unittest.TestCase):
 
         pyro.clear_param_store()
         params = model.fit(
-            data, method="MCMC", prior_fn=prior_fn, seed=42, num_steps=100
+            data, estimator="MCMC", prior_fn=prior_fn, seed=42, num_steps=100
         )
 
         self.assertIn("x1_concen1", params)
@@ -571,7 +567,7 @@ class TestFBNMethods(unittest.TestCase):
 
         params = model.fit(
             df,
-            method="SVI",
+            estimator="SVI",
             optimizer=pyro.optim.Adam({"lr": 0.05}),
             seed=42,
             num_steps=200,
@@ -721,7 +717,7 @@ class TestFBNMethods(unittest.TestCase):
 
         params = model.fit(
             df,
-            method="MCMC",
+            estimator="MCMC",
             prior_fn=prior_fn,
             seed=42,
             num_steps=100,
